move GetIdentifierTypes into separate VDServer command/task
