+++ vdc/abothe/comserver/IVDServer.cs
@@ -27,9 +27,10 @@ namespace DParserCOMServer
 		void GetSemanticExpansionsResult (out string stringList);
 		void IsBinaryOperator (string filename, uint startLine, uint startIndex, uint endLine, uint endIndex, out bool pIsOp);
 		void GetParseErrors (string filename, out string errors);
-		void GetIdentifierTypes (string filename, out string types);
 		void GetBinaryIsInLocations (string filename, out object locs); // array of pairs of DWORD
 		void GetLastMessage (out string message);
+		void GetIdentifierTypes(string filename, int startLine, int endLine, int flags);
+		void GetIdentifierTypesResult(out string types);
 		void GetDefinition (string filename, int startLine, int startIndex, int endLine, int endIndex);
 		void GetDefinitionResult (out int startLine, out int startIndex, out int endLine, out int endIndex, out string filename);
 		void GetReferences (string filename, string tok, uint line, uint idx, string expr);
